final version of rest app

+++ Distributor/demo/src/main/java/ru/paramonova/models/Circle.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Getter
 @ToString
 public class Circle {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
     private boolean color;
 }
